+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+- In development directive generators (`export`, `import`, `library`, `part`, `part of`)
+- In development script generators (`library`, `library part`)
+
 ## 0.0.1
 
 - Initial release